Moves `Param` from `utils/param.py` to `fthmc/config.py`
